requirements-test.txt
# This file is autogenerated by Prequ.  To update, run:
#
#   prequ update
#
attrs==17.4.0             # via pytest
coverage==4.5.1           # via pytest-cov
mock==2.0.0
mock-open==1.3.1
pbr==3.1.1                # via mock
pluggy==0.6.0             # via pytest
py==1.5.2                 # via pytest
pytest==3.4.2
pytest-cov==2.5.1
pytest-django==3.1.2
pyyaml==3.12
six==1.11.0               # via mock, pytest
